updated rappture's bindings and units dictionaries for compiling on 64bit platforms. previously (on our 32bit systems), we used integers as dictionary keys. This did not cause a problem because our pointers were also 32bits wide. On the 64bit platforms we test on, the integers are 32bits, but the pointers (and longs) are 64bits wide. The short term solution is to change our dictionary key to a long int, as was done in this changeset. the long term solution, if you choose to accept, is to swap out our RpDict class for something more robust like the blt hash table.

+++ trunk/src/core/RpBindingsDict.h
@@ -27,8 +27,8 @@ class RpLibrary;
 class RpUnits;
 */
 
-#define DICT_TEMPLATE_L <int,RpLibrary*>
-#define DICT_TEMPLATE_U <int,std::string>
+#define DICT_TEMPLATE_L <long,RpLibrary*>
+#define DICT_TEMPLATE_U <long,std::string>
 
 // global declaration of library and units dictionaries
 // for languages that cannot return rappture objects.
